Fixes errors in WebIDL bindings. Moves the multitap delay signal to the correct directory.
